feat(cyan-magenta): rename Checkerboard to Compare
