+++ examples/compare-image.html
@@ -22,9 +22,9 @@
 
         const fixedImage = new URL(
           // 'test-data/HeadMRVolume.nrrd',
-          'test-data/HeadMRVolumeLabels.nrrd',
+          // 'test-data/HeadMRVolumeLabels.nrrd',
           // 'test-data/HeadMRVolumeLabelsSmaller.nrrd',
-          // 'test-data/HeadMRVolume2Components.nrrd',
+          'test-data/HeadMRVolume2Components.nrrd',
           window.location.origin
         )
