+++ src/IO/Compare/.gitignore
@@ -1,0 +1,3 @@
+/emscripten-build/*
+!emscripten-build/compare*
+output.png
